requirements.txt
pillow
discord
selenium